refactor(query): move hasher to the separate package
