+++ telegram/query/hasher/hash_test.go
@@ -1,4 +1,4 @@
-package query
+package hasher
 
 import (
 	"testing"
